Trait saveload::Load

pub trait Load: Sized {
    type AuxData;
    fn load(buffer: &mut impl Read, data: &Self::AuxData) -> Result<Self>;
}

Associated Types

type AuxData

The type of the auxiliary data needed for loading.

Example

impl Load for FpVector {
    type AuxData = u32;

    fn load(buffer : &mut impl Read, p : &u32) -> io::Result<Self> {
        let dimension = usize::load(buffer, &())?;

        if dimension == 0 {
            return Ok(FpVector::new(*p, 0));
        }
        ...
    }
}

Required methods

fn load(buffer: &mut impl Read, data: &Self::AuxData) -> Result<Self>

The auxiliary data is a borrow. This is useful for Vec<T> where we pass the same auxiliary data to all the entries in the Vec.

If AuxData = (), then the load function will look like Load::load(&mut f, &()), which is a pretty common occurrence.